Move NN iterators into own package
